:recycle: split ImportTransactionsModal into smaller component files
