+++ packages/desktop-client/src/components/modals/ImportTransactionsModal/SubLabel.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { theme } from '../../../style';
+import { Text } from '../../common/Text';
+
+export function SubLabel({ title }) {
+  return (
+    <Text style={{ fontSize: 13, marginBottom: 3, color: theme.pageText }}>
+      {title}
+    </Text>
+  );
+}
