+++ packages/desktop-client/src/components/modals/ImportTransactionsModal/index.ts
@@ -1,0 +1,1 @@
+export { ImportTransactionsModal } from './ImportTransactionsModal';
