+++ upcoming-release-notes/3552.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [MatissJanis]
+---
+
+Split `ImportTransactionsModal` file in multiple smaller component files.
